Modified edly-trial.py for LMS and Studio Open edX imports its logging configuration into LOGGING settings variable in production.py using the function get_logger_config defined in openedx/core/lib/logsettings.py. The function defines a logging config that is suitable for using with a centralized syslog daemon running on Linux systems, which in the case of Ubuntu/Debian systems is rsyslog, via /dev/log virtual device.
Using its Ansible configuration, Open edX configures rsyslog to arrange
the different log directories in /edx/var/log where all logs are sent.

In a Docker container (or any jailed environment for that matter), the
device /dev/log is not available and subsequently rsyslog service is
also not available inside a container. This due to the reason that
containers have a restricted view of the system and a Docker container
is meant to run only one serivce at a time. Using the logging drivers,
Docker catches all logs sent to stdout or stderr streams.

This commit modifies the logging behavior to conform with Docker logging
capabilities and atempts to send all logs to stdout.